Move generation resource related pub structs to their own files
