+++ src/generation/debug/tile_debugger.rs
@@ -2,8 +2,8 @@ use crate::constants::*;
 use crate::coords::Point;
 use crate::coords::point::{ChunkGrid, TileGrid, World};
 use crate::events::{MouseClickEvent, RegenerateWorldEvent, ToggleDebugInfo};
-use crate::generation::lib::{ObjectComponent, Tile, TileMeshComponent};
-use crate::generation::resources::{ChunkComponentIndex, GenerationResourcesCollection};
+use crate::generation::lib::{GenerationResourcesCollection, ObjectComponent, Tile, TileMeshComponent};
+use crate::generation::resources::ChunkComponentIndex;
 use crate::resources::Settings;
 use bevy::app::{App, Plugin, Update};
 use bevy::log::*;
