+++ src/generation/lib/mod.rs
@@ -6,6 +6,7 @@ mod draft_tile;
 mod layered_plane;
 mod neighbours;
 mod plane;
+mod resources;
 pub(crate) mod shared;
 mod terrain_type;
 mod tile;
@@ -19,6 +20,7 @@ pub use draft_tile::DraftTile;
 pub use layered_plane::LayeredPlane;
 pub use neighbours::{NeighbourTile, NeighbourTiles};
 pub use plane::Plane;
+pub use resources::*;
 pub use terrain_type::TerrainType;
 pub use tile::Tile;
 pub use tile_type::TileType;
