+++ src/generation/lib/resources/mod.rs
@@ -1,0 +1,8 @@
+mod asset_collection;
+mod asset_pack;
+mod generation_resources_collection;
+mod object_resources;
+
+pub use asset_collection::AssetCollection;
+pub use asset_pack::AssetPack;
+pub use generation_resources_collection::GenerationResourcesCollection;
