+++ src/generation/object/lib/mod.rs
@@ -1,17 +1,17 @@
 mod cell;
 mod connection;
+mod iteration_result;
 mod object_data;
 mod object_grid;
 mod object_name;
 mod terrain_state;
 pub mod tile_data;
-mod iteration_result;
 
 pub use cell::Cell;
 pub use connection::Connection;
+pub use iteration_result::IterationResult;
 pub use object_data::ObjectData;
 pub use object_grid::ObjectGrid;
 pub use object_name::ObjectName;
 pub use terrain_state::TerrainState;
 pub use tile_data::TileData;
-pub use iteration_result::IterationResult;
