+++ src/generation/object/object_generator.rs
@@ -2,13 +2,10 @@ use crate::constants::*;
 use crate::coords::Point;
 use crate::coords::point::ChunkGrid;
 use crate::generation::lib::shared::CommandQueueTask;
-use crate::generation::lib::{Chunk, ObjectComponent, Tile, shared};
-use crate::generation::object::lib::ObjectName;
-use crate::generation::object::lib::tile_data::TileData;
-use crate::generation::object::lib::{ObjectData, ObjectGrid};
+use crate::generation::lib::{AssetCollection, Chunk, GenerationResourcesCollection, ObjectComponent, Tile, shared};
+use crate::generation::object::lib::{ObjectData, ObjectGrid, ObjectName, TileData};
 use crate::generation::object::wfc;
 use crate::generation::object::wfc::WfcPlugin;
-use crate::generation::resources::{AssetCollection, GenerationResourcesCollection};
 use crate::resources::Settings;
 use bevy::app::{App, Plugin, Update};
 use bevy::color::{Color, Luminance};
