+++ src/generation/world/world_generator.rs
@@ -2,8 +2,10 @@ use crate::components::{AnimationMeshComponent, AnimationTimer};
 use crate::constants::*;
 use crate::coords::Point;
 use crate::coords::point::World;
-use crate::generation::lib::{Chunk, ChunkComponent, Plane, TerrainType, Tile, TileMeshComponent, shared};
-use crate::generation::resources::{GenerationResourcesCollection, Metadata};
+use crate::generation::lib::{
+  Chunk, ChunkComponent, GenerationResourcesCollection, Plane, TerrainType, Tile, TileMeshComponent, shared,
+};
+use crate::generation::resources::Metadata;
 use crate::generation::world::post_processor;
 use crate::resources::Settings;
 use bevy::app::{App, Plugin};
